refactor: extract trainings section
